Fix A067857

+++ src/irvine/oeis/a067/A067857.java
@@ -16,7 +16,7 @@ public class A067857 extends CachedSequence {
 
   /** Construct the sequence. */
   public A067857() {
-    super(1, Integer.class, (self, n) -> H.nextQ()
+    super(1, Long.class, (self, n) -> H.nextQ()
       .subtract(Rationals.SINGLETON.sumdiv(n, k -> k.equals(n) ? Q.ZERO : new Q(self.a(k), Functions.FACTORIAL.z(k))))
       .multiply(Functions.FACTORIAL.z(n)).toZ());
   }
